Refactor and update tests
