+++ loxnetTests/ParserTests.cs
@@ -1,0 +1,18 @@
+﻿using de.softwaremess.loxnet;
+
+namespace de.softwaremess.loxnetTests
+{
+    public class ParserTests
+    {
+        [SetUp]
+        public void Setup()
+        {
+        }
+
+        [Test]
+        public void CanCreateParserClass()
+        {
+            Parser parser = new Parser();
+        }
+    }
+}
